# backend/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "MEval.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost/"
    }
  ],
  "paths": {
...
    "/api/v1/imports/dry-run": {
      "post": {
        "tags": [
          "Imports"
        ],
        "summary": "Validate and dry-run user import file, staging rows in database",
        "operationId": "DryRunImport",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "file"
                ],
                "type": "object",
                "allOf": [
                  {
                    "type": "object",
                    "properties": {
                      "file": {
                        "type": "string",
                        "format": "binary"
                      }
                    }
                  },
                  {
                    "type": "object",
                    "properties": {
                      "duplicateStrategy": {
                        "type": "integer"
                      }
                    }
                  },
                  {
                    "type": "object",
                    "properties": {
                      "commitPolicy": {
                        "type": "integer"
                      }
                    }
                  }
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/imports/{id}/execute": {
      "post": {
        "tags": [
          "Imports"
        ],
        "summary": "Execute validated import batch with live DB re-validation",
        "operationId": "ExecuteImport",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/imports/{id}/cancel": {
      "post": {
        "tags": [
          "Imports"
        ],
        "summary": "Cancel pending or validated import batch",
        "operationId": "CancelImport",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/imports/{id}/rollback": {
      "post": {
        "tags": [
          "Imports"
        ],
        "summary": "Rollback created users from an import batch",
        "operationId": "RollbackImport",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/imports/{id}": {
      "get": {
        "tags": [
          "Imports"
        ],
        "summary": "Get import batch details by ID",
        "operationId": "GetImportById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/imports/{id}/errors.xlsx": {
      "get": {
        "tags": [
          "Imports"
        ],
        "summary": "Download error report spreadsheet for import batch",
        "operationId": "GetImportErrors",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
...
}